project/requirements.txt
alembic==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.2.2
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
